1_basic_operations/genai_multimodality101/requirements.txt
openai
google.generativeai
streamlit
python-dotenv